update special note section
